Zmiana interfejsów TaskController i UserController
* Zmiana typów argumentów i wartości zwracanyc
* TaskController przyjmuje argumenty typu Task
* UserController przyjmuje argumenty typu User i zwraca User
