+++ kuba-kostrzewski/simpleToDoApp/src/main/java/pl/kostrzej/simpleToDoApp/components/user/UserController.java
@@ -2,4 +2,6 @@ package pl.kostrzej.simpleToDoApp.components.user;
 
 public interface UserController {
     void addUser();
+    User addTask(User user);
+    User getUser(long id);
 }
